refactor(ensindexer): cleanups from simplify pass
- move PonderAppShutdownManager type + isPonderAppShutdownManager guard
  to @ensnode/ponder-sdk where they sit alongside PonderAppContext, since
  they mirror a Ponder runtime type. local-ponder-context.ts now imports
  them instead of redeclaring.
- introduce AbortSignalGetter type alias in @ensnode/ponder-sdk so
  PonderClient and LocalPonderClient share one definition instead of
  three inline spellings.
- isAbortError: check error.name structurally instead of via
  instanceof Error so DOMException AbortErrors (the actual rejection
  type from fetch aborts) are caught.
- localPonderContext Proxy: ignore symbol property reads so well-known
  symbols (Symbol.toPrimitive, etc.) don't blow past the keyof cast.
- trim a duplicate staleness-contract comment in local-ponder-client.ts.
