+++ packages/ponder-sdk/src/local-ponder-client.ts
@@ -1,7 +1,7 @@
 import type { BlockNumberRangeWithStartBlock } from "./blockrange";
 import type { CachedPublicClient } from "./cached-public-client";
 import type { ChainId, ChainIdString } from "./chains";
-import { PonderClient } from "./client";
+import { type AbortSignalGetter, PonderClient } from "./client";
 import { deserializeChainId } from "./deserialize/chains";
 import {
   type ChainIndexingMetrics,
@@ -82,14 +82,14 @@ export class LocalPonderClient extends PonderClient {
    * @param ponderPublicClients All cached public clients provided by the local Ponder app
    *                            (may include non-indexed chains).
    * @param ponderAppContext The internal context of the local Ponder app.
-   * @param getAbortSignal Optional getter invoked at fetch time to attach an `AbortSignal` to outgoing HTTP requests. Use a getter (not a captured signal) when the underlying signal can change identity over time — e.g. across Ponder dev-mode hot reloads.
+   * @param getAbortSignal Optional {@link AbortSignalGetter} invoked at fetch time. See its docs for the staleness contract.
    */
   constructor(
     indexedChainIds: Set<ChainId>,
     indexedBlockranges: Map<ChainId, BlockNumberRangeWithStartBlock>,
     ponderPublicClients: Record<ChainIdString, CachedPublicClient>,
     ponderAppContext: PonderAppContext,
-    getAbortSignal?: () => AbortSignal | undefined,
+    getAbortSignal?: AbortSignalGetter,
   ) {
     super(ponderAppContext.localPonderAppUrl, getAbortSignal);
